Some great improvements in this commit include:
 * Decoupled the collection application into a separate lists app;
 * Added more permissions to the models;
 * Initial (CRUD) list support;
 * Messaging (confirmation,...) support, with templates and styles;
 * Implemented list navigation item;
 * Improved JS (focus on load, focus management, messages,...);
 * Added 'small' and 'negative' buttons to the styles;
 * Converted Named URLs to namespaces;

Todo now, before deployment of a first demo:

 * Movie details page;
 * Add to/remove from lists;
 * Browse per genre/actor;
 * Add some user utilities such as password change/restore, email address support,...

Would be nice to create a vBulletin postcommit hook for github which automatically
posts commit messages to a topic on the AnimeDB forum. I'll look into this later on.

Permissions usage should be improved too; each view should use a different permission
(or more, depending on the action) and fixtures with user groups should be created
(with groups such as 'Basic', 'Premium' and so on for the different services).

The database importing tool should also be improved and fixtures uploaded to
the github downloads section to be available later on.